added empty component in fav page and changed icons and added styles in header
