+++ src/pages/Favorites/favorites.scss
@@ -194,3 +194,18 @@
     }
   }
 }
+
+.fav-empty {
+  padding-top: 20px;
+  @include flex(column, center, center);
+  &-text {
+      font-family: "Roboto", sans-serif;
+      font-size: 14px;
+      @include breakpoint-mf("sm") {
+          font-size: 16px;
+      }
+      @include breakpoint-mf("md") {
+          font-size: 18px;
+      }
+  }
+}
